only ask whether to remove text output if there is leftover text output

+++ entry.sh
@@ -14,14 +14,18 @@ taskrange_end="$2"
 job_name="$3"
 
 
-echo "Shall I clear the folders with output and error text? [y/n]"
-read answer
+# Clear text output directory?
+if [ -d "./textoutput/$job_name" ]; then
+    echo "Shall I clear the folders with output and error text? [y/n]"
+    read answer
 
-if [ "$answer" == "y" ] || [ "$answer" == "" ]; then
-	echo "Clearing.."
-	rm -rf ./textoutput/"$job_name"
+    if [ "$answer" == "y" ] || [ "$answer" == "" ]; then
+        echo "Clearing.."
+        rm -rf ./textoutput/"$job_name"
+    fi
 fi
 
+# if textoutput directory doesn't exist, create
 if [ ! -d "./textoutput/$job_name" ]; then
 	echo "Creating textoutput directory textoutput"
 	mkdir -p ./textoutput/"$job_name"
